refactor: make cli commands independent of RPC
Create a query interface for data required by the RPC and commands layer
to be used by both of them. Move as much code from 'node' to library
creates as possible without violating GPLv3.
