+++ pallets/session-validator-management/rpc/Cargo.toml
@@ -4,31 +4,7 @@ version = "0.0.1"
 edition = "2021"
 
 [dependencies]
-parity-scale-codec = { workspace = true }
-scale-info = { workspace = true }
-hex = { workspace = true }
-jsonrpsee = { workspace = true }
-serde = { workspace = true }
 derive-new = { workspace = true }
-
-sp-core = { workspace = true }
-sp-runtime = { workspace = true }
-sp-api = { workspace = true }
-frame-support = { workspace = true }
-sp-blockchain = { workspace = true }
-main-chain-follower-api = { workspace = true, features = ['std', 'candidate-source'] }
-sp-session-validator-management = { workspace = true, default-features = true }
-sp-sidechain = { workspace = true, default-features = true }
+jsonrpsee = { workspace = true }
 sidechain-domain = { workspace = true, default-features = true }
-sidechain-block-search = { workspace = true }
-authority-selection-inherents = { workspace = true, default-features = true }
-thiserror = { workspace = true }
-plutus = { workspace = true }
-
-[dev-dependencies]
-serde_json = { workspace = true }
-sp-std = { workspace = true }
-tokio = { workspace = true, features = ["macros", "time", "parking_lot"] }
-hex-literal = { workspace = true }
-minicbor = { workspace = true }
-plutus-datum-derive = { workspace = true }
+sp-session-validator-management-query = { workspace = true, default-features = true }
