+++ pallets/session-validator-management/rpc/src/lib.rs
@@ -1,8 +1,3 @@
-pub mod get_registrations;
-pub mod types;
-
-use authority_selection_inherents::authority_selection_inputs::AuthoritySelectionInputs;
-use authority_selection_inherents::filter_invalid_candidates::CandidateValidationApi;
 use derive_new::new;
 use jsonrpsee::{
 	core::async_trait,
@@ -10,23 +5,11 @@ use jsonrpsee::{
 	proc_macros::rpc,
 	types::{ErrorObject, ErrorObjectOwned},
 };
-use main_chain_follower_api::candidate::RawPermissionedCandidateData;
-use main_chain_follower_api::CandidateDataSource;
-use plutus::ToDatum;
-use sidechain_block_search::{predicates::AnyBlockInEpoch, FindSidechainBlock, SidechainInfo};
-use sidechain_domain::{MainchainPublicKey, McEpochNumber, ScEpochNumber};
-use sp_api::ProvideRuntimeApi;
-use sp_blockchain::{HeaderBackend, Info};
-use sp_core::bytes::to_hex;
-use sp_runtime::traits::NumberFor;
-use sp_runtime::traits::{Block as BlockT, Zero};
-use sp_session_validator_management::SessionValidatorManagementApi;
-use sp_sidechain::{GetSidechainParams, GetSidechainStatus};
+use sidechain_domain::{MainchainPublicKey, McEpochNumber};
+use sp_session_validator_management_query::types::*;
+use sp_session_validator_management_query::SessionValidatorManagementQueryApi;
 use std::sync::Arc;
-use types::*;
 
-#[cfg(test)]
-mod tests;
 #[rpc(client, server, namespace = "sidechain")]
 pub trait SessionValidatorManagementRpcApi {
 	/// Returns the committee for given sidechain epoch. The order of the list represents the order of slot allocation.
@@ -53,106 +36,17 @@ pub trait SessionValidatorManagementRpcApi {
 }
 
 #[derive(new)]
-pub struct SessionValidatorManagementRpc<
-	C,
-	Block,
-	SessionKeys: parity_scale_codec::Decode,
-	CrossChainPublic,
-	SidechainParams: parity_scale_codec::Decode + ToDatum + Clone + Send + Sync + 'static,
-> {
-	client: Arc<C>,
-	candidate_data_source: Arc<dyn CandidateDataSource + Send + Sync>,
-	_marker: std::marker::PhantomData<(Block, SessionKeys, CrossChainPublic, SidechainParams)>,
+pub struct SessionValidatorManagementRpc<T> {
+	query_api: Arc<T>,
 }
 
 #[async_trait]
-impl<
-		C,
-		Block,
-		SessionKeys: parity_scale_codec::Decode + Send + Sync + 'static,
-		SidechainParams: parity_scale_codec::Decode + ToDatum + Clone + Send + Sync + 'static,
-		CrossChainPublic: parity_scale_codec::Decode
-			+ parity_scale_codec::Encode
-			+ AsRef<[u8]>
-			+ Send
-			+ Sync
-			+ 'static,
-	> SessionValidatorManagementRpcApiServer
-	for SessionValidatorManagementRpc<C, Block, SessionKeys, CrossChainPublic, SidechainParams>
+impl<T> SessionValidatorManagementRpcApiServer for SessionValidatorManagementRpc<T>
 where
-	Block: BlockT,
-	NumberFor<Block>: From<u32> + Into<u32>,
-	C: Send + Sync + 'static,
-	C: ProvideRuntimeApi<Block>,
-	C: HeaderBackend<Block>,
-	C::Api: sp_api::Core<Block>,
-	C::Api: GetSidechainStatus<Block>,
-	C::Api: SessionValidatorManagementApi<
-		Block,
-		SessionKeys,
-		CrossChainPublic,
-		AuthoritySelectionInputs,
-		ScEpochNumber,
-	>,
-	C::Api: GetSidechainParams<Block, SidechainParams>,
-	C::Api: CandidateValidationApi<Block>,
+	T: SessionValidatorManagementQueryApi + Send + Sync + 'static,
 {
 	fn get_epoch_committee(&self, epoch_number: u64) -> RpcResult<GetCommitteeResponse> {
-		let epoch_number = ScEpochNumber(epoch_number);
-		let Info { genesis_hash, best_number: latest_block, best_hash, .. } = self.client.info();
-
-		if epoch_number.is_zero() {
-			let (_, genesis_committee) = (self.client.runtime_api())
-				.get_current_committee(genesis_hash)
-				.map_err(error_object_from)?;
-			return Ok(GetCommitteeResponse::new(epoch_number, genesis_committee));
-		}
-
-		let first_epoch = {
-			let second_block = (self.client)
-				.hash(1.into())
-				.map_err(error_object_from)?
-				.ok_or(error_object_from("Only the Genesis Block exists at the moment!"))?;
-			(self.client.runtime_api())
-				.get_sidechain_status(second_block)
-				.map_err(error_object_from)?
-				.epoch
-		};
-
-		if epoch_number < first_epoch {
-			return Err(error_object_from_str(format!(
-				"Epoch {} is earlier than the Initial Epoch!",
-				epoch_number
-			)));
-		}
-
-		let epoch_of_latest_block =
-			self.client.get_epoch_of_block(latest_block).map_err(error_object_from)?;
-
-		if epoch_number > epoch_of_latest_block.next() {
-			return Err(error_object_from_str(format!(
-				"Committee is unknown for epoch {epoch_number}"
-			)));
-		}
-
-		let (_, committee) = if epoch_number == epoch_of_latest_block.next() {
-			(self.client.runtime_api())
-				.get_next_committee(best_hash)
-				.map_err(error_object_from)?
-				.ok_or(error_object_from_str(format!(
-					"Committee is unknown for the next epoch: {epoch_number}"
-				)))
-		} else {
-			let block_hash = self
-				.client
-				.find_block(AnyBlockInEpoch { epoch: epoch_number })
-				.map_err(error_object_from)?;
-			(self.client.runtime_api())
-				.get_current_committee(block_hash)
-				.map_err(error_object_from)
-		}?;
-
-		Ok(GetCommitteeResponse::new(epoch_number, committee))
+		self.query_api.get_epoch_committee(epoch_number).map_err(error_object_from_str)
 	}
 
 	async fn get_registrations(
@@ -160,71 +54,23 @@ where
 		mc_epoch_number: McEpochNumber,
 		mc_public_key: MainchainPublicKey,
 	) -> RpcResult<Vec<CandidateRegistrationEntry>> {
-		let api = self.client.runtime_api();
-		let best_block = self.client.info().best_hash;
-		let scripts = api.get_main_chain_scripts(best_block).map_err(error_object_from)?;
-		let mut registrations_map = self
-			.candidates_registrations_for_epoch(
-				mc_epoch_number,
-				scripts.committee_candidate_address,
-			)
-			.await?;
-		Ok(registrations_map.remove(&to_hex(&mc_public_key.0, false)).unwrap_or(vec![]))
+		self.query_api
+			.get_registrations(mc_epoch_number, mc_public_key)
+			.await
+			.map_err(error_object_from_str)
 	}
 
 	async fn get_ariadne_parameters(
 		&self,
 		epoch_number: McEpochNumber,
 	) -> RpcResult<AriadneParameters> {
-		let api = self.client.runtime_api();
-		let best_block = self.client.info().best_hash;
-		let scripts = api.get_main_chain_scripts(best_block).map_err(error_object_from)?;
-		let ariadne_parameters_response = self
-			.candidate_data_source
-			.get_ariadne_parameters(
-				epoch_number,
-				scripts.d_parameter_policy,
-				scripts.permissioned_candidates_policy,
-			)
+		self.query_api
+			.get_ariadne_parameters(epoch_number)
 			.await
-			.map_err(error_object_from)?;
-
-		let candidate_registrations = self
-			.candidates_registrations_for_epoch(epoch_number, scripts.committee_candidate_address)
-			.await?;
-		let validate_permissioned_candidate = |candidate: &RawPermissionedCandidateData| {
-			api.validate_permissioned_candidate_data(
-				best_block,
-				sidechain_domain::PermissionedCandidateData {
-					sidechain_public_key: candidate.sidechain_public_key.clone(),
-					aura_public_key: candidate.aura_public_key.clone(),
-					grandpa_public_key: candidate.grandpa_public_key.clone(),
-				},
-			)
-		};
-
-		Ok(AriadneParameters {
-			d_parameter: ariadne_parameters_response.d_parameter.into(),
-			permissioned_candidates: ariadne_parameters_response
-				.permissioned_candidates
-				.into_iter()
-				.map(|candidate| {
-					let validation_result =
-						validate_permissioned_candidate(&candidate).map_err(error_object_from)?;
-					Ok::<PermissionedCandidateData, ErrorObjectOwned>(
-						PermissionedCandidateData::new(candidate, validation_result),
-					)
-				})
-				.collect::<Result<Vec<_>, _>>()?,
-			candidate_registrations,
-		})
+			.map_err(error_object_from_str)
 	}
 }
 
-pub fn error_object_from<T: std::fmt::Debug>(err: T) -> ErrorObjectOwned {
-	ErrorObject::owned::<u8>(-1, format!("{err:?}"), None)
-}
-
-pub fn error_object_from_str(msg: impl Into<String>) -> ErrorObjectOwned {
+fn error_object_from_str(msg: impl Into<String>) -> ErrorObjectOwned {
 	ErrorObject::owned::<u8>(-1, msg, None)
 }
